refactor(hooks): extract the banned-terms publish gate from hook_router (#35, #2384)
Wave-2 PR2 of the hook_router split. Pure behavior-preserving move of the
PreToolUse banned-terms publish gate into hooks/scripts/banned_terms_gate.py,
re-exported into the _HANDLERS dispatch table so the gate decision, the deny
payload, and the existing banned-terms test corpus are all unchanged.

The five H-cluster symbols move verbatim: _banned_terms_gate_enabled,
handle_banned_terms_pretool, _BANNED_TERMS_CREDENTIAL_DENY,
_banned_term_marker_blocks, _run_banned_terms_pretool. Module-top imports are
stdlib plus already-extracted hooks/scripts siblings only (cold-import safe for
the bare python3 hook subprocess); the shared spine helpers emit_pretooluse_deny
and _resolve_cwd_repo stay in the router and are back-imported lazily inside the
handler bodies, the hooks/scripts sibling back-import the import-direction fitness
test permits.

hook_router shrinks 8119 -> 7997 LOC (module-health green); new sibling 161 LOC /
1 public function. A cold-import smoke assertion for the sibling is added to
test_hook_scripts_py_version_compat.py, mirroring PR1.

Refs #2384.
